all: migrate from double to float data type for all computation

+++ src/audio/AudioSource.hpp
@@ -8,7 +8,7 @@ namespace Audio {
 class AudioSource {
   public:
     virtual ~AudioSource() {}
-    virtual void read(std::vector<double> &samples) = 0;
+    virtual void read(std::vector<float> &samples) = 0;
     virtual unsigned int getSampleRate() = 0;
 };
