+++ src/main/Configuration.hpp
@@ -24,26 +24,26 @@ struct Settings {
     unsigned int dftSize = 1024;
     RealDft::WindowFunction dftWf = RealDft::WindowFunction::Hann;
     /* Spectrogram Settings */
-    double magnitudeMin = 0.0;
-    double magnitudeMax = 45.0;
+    float magnitudeMin = 0.0;
+    float magnitudeMax = 45.0;
     bool magnitudeLog = true;
     SpectrumRenderer::ColorScheme colors = SpectrumRenderer::ColorScheme::Heat;
     /* Initial settings when switching between logarithmic/linear in UI */
-    double magnitudeLogMin = 0.0;
-    double magnitudeLogMax = 50.0;
-    double magnitudeLinearMin = 0.0;
-    double magnitudeLinearMax = 50.0;
+    float magnitudeLogMin = 0.0;
+    float magnitudeLogMax = 50.0;
+    float magnitudeLinearMin = 0.0;
+    float magnitudeLinearMax = 50.0;
 };
 
 struct Limits {
     /* Linear magnitude min, max, step */
-    double magnitudeLinearMin = 0.0;
-    double magnitudeLinearMax = 1000.0;
-    double magnitudeLinearStep = 25.0;
+    float magnitudeLinearMin = 0.0;
+    float magnitudeLinearMax = 1000.0;
+    float magnitudeLinearStep = 25.0;
     /* Logarithmic magnitude min, max, step */
-    double magnitudeLogMin = -80.0;
-    double magnitudeLogMax = 80.0;
-    double magnitudeLogStep = 5.0;
+    float magnitudeLogMin = -80.0;
+    float magnitudeLogMax = 80.0;
+    float magnitudeLogStep = 5.0;
     /* DFT size min, max */
     unsigned int dftSizeMin = 64;
     unsigned int dftSizeMax = 8192;
